Rename poor meaning globals.py to context.py
